requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.13
pydantic>=2.5.2
python-dotenv>=1.0.0
openai>=1.3.0
numpy>=1.24.3
pandas>=2.0.3
tiktoken>=0.5.1
scikit-learn>=1.3.0
matplotlib>=3.7.2
nltk>=3.8.1
spacy>=3.7.2
fastapi>=0.104.1
uvicorn>=0.24.0 